feat: Add new functions for chat completions and image generation
Add new functions for chat completions and image generation in the  package.

The  struct now includes the following functions:

1. : Handles chat completions by sending a POST request to the endpoint .

2. : Handles text completions by sending a POST request to the endpoint .

3. : Handles image generation by sending a POST request to the endpoint .

4. : Handles image generation with base64 encoded JSON by sending a POST request to the endpoint .

5. : Handles image variations by sending a POST request
